[1118] Update location in Support
