+++ app/controllers/concerns/success_message.rb
@@ -9,8 +9,8 @@ module SuccessMessage
     flash[:success] = I18n.t('success.saved', value:)
   end
 
-  def schools_added_message(locations)
-    items_added = locations.size > 1 ? "#{locations.size} locations" : '1 location'
+  def schools_added_message(schools)
+    items_added = schools.size > 1 ? "#{schools.size} schools" : '1 school'
     flash[:success] = I18n.t('success.added', items_added:)
   end
 end
