+++ app/views/layouts/provider_record.html.erb
@@ -14,7 +14,7 @@
     { name: "Details", url: support_recruitment_cycle_provider_path(@provider.recruitment_cycle_year, @provider) },
     { name: "Users", url: support_recruitment_cycle_provider_users_path(@provider.recruitment_cycle_year, @provider) },
     { name: "Courses", url: support_recruitment_cycle_provider_courses_path(@provider.recruitment_cycle_year, @provider) },
-    { name: "Locations", url: support_recruitment_cycle_provider_schools_path(@provider.recruitment_cycle_year, @provider) }
+    { name: "Schools", url: support_recruitment_cycle_provider_schools_path(@provider.recruitment_cycle_year, @provider) }
   ]) %>
 
   <%= yield %>
