+++ app/views/publish/providers/schools/_school.html.erb
@@ -1,4 +1,4 @@
-<tr class="govuk-table__row location-row">
+<tr class="govuk-table__row school-row">
   <th class="govuk-table__header name" scope="row">
     <%= govuk_link_to school.location_name, edit_publish_provider_recruitment_cycle_school_path(@provider.provider_code, school.recruitment_cycle.year, school.id) %>
   </th>
