--- app/views/support/locations/new.html.erb
@@ -1,26 +1,0 @@
-<%= render PageTitle.new(title: "support.providers.locations.new") %>
-
-<h1 class="govuk-heading-l">
-  <span class="govuk-caption-l"><%= "#{@provider.provider_name} (#{@provider.provider_code})" %></span>
-    Add location
-</h1>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds-from-desktop">
-    <%= render partial: "support/locations/form", locals: { site: @site, form_url: support_recruitment_cycle_provider_locations_path } %>
-
-    <p class="govuk-body">
-      <%= govuk_link_to(t("cancel"), support_recruitment_cycle_provider_locations_path) %>
-    </p>
-
-    <% if @site.persisted? %>
-      <p class="govuk-body">
-        <%= govuk_button_to t("support.actions.delete", resource: "location"),
-            support_provider_location_path(@provider, @site),
-            method: :delete,
-            class: "govuk-button govuk-button--warning" %>
-      </p>
-    <% end %>
-
-  </div>
-</div>
