--- app/views/support/providers/locations_check/show.html.erb
@@ -1,39 +1,0 @@
-<%= render PageTitle.new(title: "support.providers.locations.check") %>
-<% content_for :before_content do %>
-  <%= govuk_back_link_to(new_support_recruitment_cycle_provider_location_path) %>
-<% end %>
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds-from-desktop">
-    <%= form_with(model: @location_form, url: support_recruitment_cycle_provider_check_location_path(@provider.recruitment_cycle_year, @provider), method: :put, local: true) do |f| %>
-
-    <fieldset class="govuk-fieldset">
-      <legend class="govuk-fieldset__legend govuk-fieldset__legend--l"> <span class="govuk-caption-l"><%= t("support.providers.locations.new") %> - <%= @provider.provider_name %> </span>
-        <h1 class="govuk-fieldset__heading govuk-!-margin-bottom-3"><%= t("support.providers.users.check") %></h1>
-      </legend>
-
-      <%= render GovukComponent::SummaryListComponent.new do |component|
-            component.with_row do |row|
-              row.with_key { t("support.providers.locations.location_name") }
-              row.with_value(text:  @location_form.location_name)
-              row.with_action(text: t("change"), href: new_support_recruitment_cycle_provider_location_path(@provider.recruitment_cycle_year, @provider), classes: "location_name", visually_hidden_text: t("support.providers.locations.location_name"))
-            end
-
-            component.with_row do |row|
-              row.with_key { t("support.providers.locations.urn") }
-              row.with_value(text: @location_form.urn)
-              row.with_action(text: t("change"), href: new_support_recruitment_cycle_provider_location_path(@provider.recruitment_cycle_year, @provider), classes: "urn", visually_hidden_text: t("support.providers.locations.urn"))
-            end
-
-            component.with_row do |row|
-              row.with_key { t("support.providers.locations.address") }
-              row.with_value(text: @location_form.full_address)
-              row.with_action(text: t("change"), href: new_support_recruitment_cycle_provider_location_path(@provider.recruitment_cycle_year, @provider), classes: "address", visually_hidden_text: t("support.providers.locations.address"))
-            end
-          end %>
-    </fieldset>
-      <%= f.govuk_submit(t("support.providers.locations.new")) %>
-      <%= f.govuk_submit(t("support.providers.locations.another"), name: "another", class: "govuk-button govuk-button--secondary govuk") %>
-    <% end %>
-    <%= govuk_link_to(t("cancel"), support_recruitment_cycle_provider_locations_path) %>
-  </div>
-</div>
