+++ app/views/support/providers/schools/check_multiple/show.html.erb
@@ -1,0 +1,28 @@
+<% page_title = t("support.providers.multiple_schools.check") %>
+<%= content_for :page_title, page_title %>
+
+<% content_for :before_content do %>
+  <%= govuk_back_link_to(support_recruitment_cycle_provider_schools_multiple_new_path(position: @school_details.count)) %>
+<% end %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds-from-desktop">
+    <%= form_with(model: @parsed_csv_school_form, url: support_recruitment_cycle_provider_schools_multiple_check_path(provider_code: @provider.provider_code, recruitment_cycle_year: @provider.recruitment_cycle_year), method: :put, local: true) do |f| %>
+      <legend class="govuk-fieldset__legend govuk-fieldset__legend--l"> <span class="govuk-caption-l"><%= "Add schools - #{@provider.decorate.name_and_code}" %></span>
+        <h1 class="govuk-fieldset__heading govuk-!-margin-bottom-3"><%= t("support.providers.multiple_schools.check") %> </h1> </legend>
+        <% @school_details.each.with_index(1) do |school_detail, position| %>
+          <%= govuk_summary_card(title: "School #{position}") do |card|
+                card.with_action { govuk_link_to("Change", support_recruitment_cycle_provider_schools_multiple_new_path(position:, goto_confirmation: true)) }
+                card.with_summary_list(rows: [{ key: { text: "Name" }, value: { text: school_detail.location_name } },
+                                              { key: { text: "Unique reference number (URN)" }, value: { classes: ("govuk-hint" if school_detail.urn.blank?), text: (school_detail.urn.presence || "Not entered") } },
+                                              { key: { text: "Address" }, value: { text: school_detail.decorate.full_address_on_seperate_lines } }])
+              end %>
+        <% end %>
+      <%= f.govuk_submit("Add schools") %>
+    <% end %>
+
+    <p class="govuk-body">
+      <%= govuk_link_to(t("cancel"), support_recruitment_cycle_provider_schools_path) %>
+    </p>
+  </div>
+</div>
