+++ config/locales/en.yml
@@ -143,10 +143,10 @@ en:
           courses:
             index: "Courses"
             edit: "Edit course details"
-          locations:
-            index: "Locations"
-            new: "Add location"
-            edit: "Edit location details"
+          schools:
+            index: "Schools"
+            new: "Add school"
+            edit: "Edit school details"
             check: "Check your answers"
           users:
             new: "Add user"
@@ -218,11 +218,11 @@ en:
           name: "Only users affiliated with a provider"
           description: "The list of all users affiliated with a provider from current recruitment cycle with columns: provider_code, provider_name, provider_type, first_name, last_name, email_address"
     providers:
-      locations:
-        new: "Add location"
-        added: "Location added"
-        another: "Save location and add another"
-        location_name: "Location name"
+      schools:
+        new: "Add school"
+        added: "School added"
+        another: "Save school and add another"
+        location_name: "School name"
         urn: "Unique reference number (URN)"
         address1: "Address line 1"
         address2: "Address line 2"
@@ -236,9 +236,9 @@ en:
         new: "Add user"
         edit: "Update user"
         check: "Check your answers"
-      multiple_locations:
-        new: "Add locations"
-        show: "Add location (%{position} of %{max})"
+      multiple_schools:
+        new: "Add schools"
+        show: "Add school (%{position} of %{max})"
         check: "Check your answers"
     flash:
       created: "%{resource} successfully created"
@@ -277,8 +277,8 @@ en:
         site:
           attributes:
             location_name:
-              blank: "Enter a location name"
-              taken: "is in use by another location"
+              blank: "Enter a school name"
+              taken: "is in use by another school"
             address1:
               blank: "Enter address line 1"
             address3:
@@ -325,7 +325,7 @@ en:
             applications_open_from:
               blank: "^Select an applications open date"
             base:
-              duplicate: "This course already exists. You should add further locations for this course to the existing profile in Publish"
+              duplicate: "This course already exists. You should add further schools for this course to the existing profile in Publish"
               visa_sponsorship_not_publishable: "Select if visas can be sponsored"
               provider_ukprn_not_publishable: "Enter a UK Provider Reference Number (UKPRN)"
               provider_ukprn_and_urn_not_publishable: "Enter a UK Provider Reference Number (UKPRN) and URN"
@@ -463,7 +463,7 @@ en:
               blank: "Enter a first name"
             last_name:
               blank: "Enter a last name"
-        support/location_form:
+        support/school_form:
           attributes:
             location_name:
               blank: "Enter a name"
@@ -478,7 +478,7 @@ en:
             urn:
               format: "Site URN must be 5 or 6 numbers"
         support/raw_csv_schools_form:
-          blank: "Enter location details"
+          blank: "Enter school details"
         publish/course_information_form:
           attributes:
             about_course:
