--- spec/support/page_objects/support/provider/locations_index.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module PageObjects
-  module Support
-    module Provider
-      class LocationsIndex < PageObjects::Base
-        set_url '/support/{recruitment_cycle_year}/providers/{provider_id}/locations'
-
-        element :add_location, '.govuk-button', text: 'Add location'
-
-        sections :locations, Sections::School, '.location-row'
-      end
-    end
-  end
-end
