+++ spec/support/page_objects/support/provider/schools_index.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module PageObjects
+  module Support
+    module Provider
+      class SchoolsIndex < PageObjects::Base
+        set_url '/support/{recruitment_cycle_year}/providers/{provider_id}/schools'
+
+        element :add_school, '.govuk-button', text: 'Add school'
+
+        sections :schools, Sections::School, '.school-row'
+      end
+    end
+  end
+end
